+++ src/main/java/ru/textanalysis/tawt/ms/internal/TypeForms.java
@@ -1,0 +1,18 @@
+package ru.textanalysis.tawt.ms.internal;
+
+public enum TypeForms {
+    UNFAMILIAR(1),
+    INITIAL(2),
+    NUMBER(3),
+    WORD(4);
+
+    private final int id;
+
+    TypeForms(int id) {
+        this.id = id;
+    }
+
+    public int getId() {
+        return id;
+    }
+}