+++ src/main/java/ru/textanalysis/tawt/ms/internal/form/GetCharacteristics.java
@@ -1,5 +1,7 @@
 package ru.textanalysis.tawt.ms.internal.form;
 
+import ru.textanalysis.tawt.ms.internal.TypeForms;
+
 public interface GetCharacteristics {
     byte getTypeOfSpeech();
 
@@ -10,4 +12,6 @@ public interface GetCharacteristics {
     boolean isInitialForm();
 
     GetCharacteristics getInitialForm();
+
+     TypeForms isTypeForm();
 }